Moving treeannotator (pre Beastx) to oldtreeannotator
